5758 - Extract common getLinkValidationMessage & buildVisitedLinksByKey
